Refactor: change customer id variable type

+++ src/main/java/com/example/customerrelationmanager/infrastructure/entity/Customer.java
@@ -17,7 +17,7 @@ import java.util.List;
 public class Customer {
 
     @Id
-    private long customerId;
+    private int customerId;
 
     private String firstName;
